Add $link object to the begin callback
During handleClick, send $link to the begin callback so the caller can evaluate the link and choose to do something else. If the return value is ignored (undefined), or true, the call will proceed. If the return value is false, the call will terminate. This is useful when the nav is reloaded in a master page and the current URL doesn't match the link URL. Sample implementation:

  $('.main-nav-list').onePageNav({
    scrollThreshold: 0.2, // Adjust if Navigation highlights too early or too late
    scrollOffset: 75, //Height of Navigation Bar,
    changeHash: true,
    begin: function ($link) {
        var dest = $link[0].href.split("#")[0];
        var curr = location.href.replace(location.hash, "");
        if (dest != curr) {
            location.href = $link[0].href;
            return false;
        }
    }
  });